requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.6.1
lightgbm==4.6.0
xgboost==2.1.4
catboost==1.2.8
scipy==1.14.1
matplotlib==3.9.2
statsmodels==0.14.4
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
mlflow==2.20.1
prefect==2.19.9
evidently==0.6.7
pydantic==2.10.6
python-dotenv==1.0.1
prometheus-client==0.21.1
pytest==8.3.4
pytest-cov==6.0.0
black==24.10.0
ruff==0.9.2
mypy==1.14.1
requests==2.32.3
pyarrow==18.1.0
PyYAML==6.0.2
streamlit==1.42.2
altair==5.5.0
